Revert move annotations.
